% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_dayslost.R
\name{calc_mean_dayslost}
\alias{calc_mean_dayslost}
\title{Calculate the mean days lost}
\usage{
calc_mean_dayslost(injd, by = NULL, overall = TRUE)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see
\code{\link[=prepare_all]{prepare_all()}}).}

\item{by}{Character specifying the name of the column according to which
compute (sports-related case) summary statistics. It should refer to a
(categorical) variable that describes the "type of case" (e.g. type of
injury). Optional, defaults to \code{NULL}.}

\item{overall}{Logical, whether to calculate overall (for all the cohort) or
athlete-wise summary statistic (i.e. number of cases). Defaults to
\code{TRUE}.}
}
\value{
The mean of the days lost. Either a numeric value (if overall
\code{TRUE}) or a data frame indicating the mean days lost per athlete.
}
\description{
Calculate the mean of the days lost due to a sports-related health problem
(e.g. disease, injury) in a cohort.
}
\examples{
calc_mean_dayslost(injd)
calc_mean_dayslost(injd, overall = FALSE)
calc_mean_dayslost(injd, by = "injury_type")
}
