% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ncases.R
\name{calc_ncases}
\alias{calc_ncases}
\title{Calculate number of cases}
\usage{
calc_ncases(injd, by = NULL, overall = TRUE)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see
\code{\link[=prepare_all]{prepare_all()}}).}

\item{by}{Character specifying the name of the column according to which
compute (sports-related case) summary statistics. It should refer to a
(categorical) variable that describes the "type of case" (e.g. type of
injury). Optional, defaults to \code{NULL}.}

\item{overall}{Logical, whether to calculate overall (for all the cohort) or
athlete-wise summary statistic (i.e. number of cases). Defaults to
\code{TRUE}.}
}
\value{
The number of cases. Either a numeric value (if overall \code{TRUE})
or a data frame indicating the number of cases per athlete.
}
\description{
Calculate the number of sports-related cases (e.g. injuries) that occurred in
a cohort during a period.
}
\examples{
calc_ncases(injd)
calc_ncases(injd, overall = FALSE)
calc_ncases(injd, by = "injury_type")
}
