% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prevalence.R
\name{calc_prevalence}
\alias{calc_prevalence}
\title{Calculate prevalence proportion}
\usage{
calc_prevalence(injd, time_period = c("monthly", "season"), by = NULL)
}
\arguments{
\item{injd}{Prepared data. An \code{injd} object.}

\item{time_period}{Character. One of "monthly" or "season", specifying the
periodicity according to which to calculate the proportions of available
and injured athletes.}

\item{by}{Character specifying the name of the column on the basis of which
to classify the injuries and calculate proportions of the injured athletes.
Defaults to \code{NULL}.}
}
\value{
A data frame containing one row for each combination of season, month
(optionally) and injury type (if \code{by} not specified, then this variable has
two categories: \emph{Available} and \emph{Injured}). Plus, three more columns,
specifying the proportion of athletes (\code{prop}) satisfying the corresponding
row's combination of values, i.e. prevalence, how many athletes were injured
at that moment with the type of injury of the corresponding row (\code{n}), over
how many athletes were at that time in the cohort (\code{n_athlete}). See Note
section.
}
\description{
Calculate the prevalence proportion of injured athletes and the proportion of
non-injured (available) athletes in the cohort, on a monthly or season basis.
Further information on the type of injury may be specified so that the
injury-specific prevalences are reported according to this variable.
}
\note{
If \code{by} is specified (and not \code{NULL}), it may happen that an athlete in
one month suffers two different types of injuries. For example, a muscle and
a ligament injury. In this case, this two injuries contribute to the
proportions of muscle and ligament injuries for that month, resulting in an
overall proportion that exceeds 100\%. Besides, the athletes in Available
category are those that did not suffer any injury in that moment
(season-month), that is, they were healthy all the time that the period
lasted.
}
\examples{
\donttest{
df_exposures <- prepare_exp(raw_df_exposures, person_id = "player_name",
                            date = "year", time_expo = "minutes_played")
df_injuries  <- prepare_inj(raw_df_injuries, person_id = "player_name",
                            date_injured = "from", date_recovered = "until")
injd         <- prepare_all(data_exposures = df_exposures,
                            data_injuries  = df_injuries,
                            exp_unit = "matches_minutes")
}

calc_prevalence(injd, time_period = "monthly", by = "injury_type")
calc_prevalence(injd, time_period = "monthly")
calc_prevalence(injd, time_period = "season", by = "injury_type")
calc_prevalence(injd, time_period = "season")
}
\references{
Bahr R, Clarsen B, Derman W, et al. International Olympic
Committee consensus statement: methods for recording and reporting of
epidemiological data on injury and illness in sport 2020 (including STROBE
Extension for Sport Injury and Illness Surveillance (STROBE-SIIS))
\emph{British Journal of Sports Medicine} 2020; 54:372-389.

Nielsen RO, Debes-Kristensen K, Hulme A, et al. Are prevalence measures
better than incidence measures in sports injury research? \emph{British
Journal of Sports Medicine} 2019; 54:396-397.
}
