% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_injfollowup}
\alias{check_injfollowup}
\title{Check the dates in injury data and exposure data}
\usage{
check_injfollowup(followup_df, data_injuries)
}
\arguments{
\item{followup_df}{Data frame created inside prepare_all() function.}

\item{data_injuries}{Data frame given in prepare_all() function.}
}
\value{
Same injury data or cut injury data in which the dates of injury
occurrences are within the dates of each person_id's follow-up period.
}
\description{
Check that dates of injury data are within the dates of exposure data.
}
\keyword{internal}
