% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{data_followup}
\alias{data_followup}
\title{Build follow-up data frame}
\usage{
data_followup(data_exposures)
}
\arguments{
\item{data_exposures}{Exposure data frame with standardized column names, in
the same fashion that \code{prepare_exp()} returns.}
}
\value{
A data frame in which each row corresponds to a player and his/her
first date (\code{t0}) and last date (\code{tf}) observed.
}
\description{
Build follow-up data frame
}
\keyword{internal}
