% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{data_injuarieslong}
\alias{data_injuarieslong}
\title{Transform injury data into a long format}
\usage{
data_injuarieslong(data_injuries)
}
\arguments{
\item{data_injuries}{Injury data frame with standardized column names, in the
same fashion that \code{prepare_inj()} returns.}
}
\value{
The \code{data_injuries} data frame in long format in which each row
corresponds to player-event.
}
\description{
Transform injury data into a long format
}
\keyword{internal}
