% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date2season}
\alias{date2season}
\title{Get the season}
\usage{
date2season(date)
}
\arguments{
\item{date}{a \link[base:Dates]{Date} object.}
}
\value{
Character specifying the respective competition season given the
date.
}
\description{
Get the season given the date.
}
\examples{
date <- Sys.Date()
date2season(date)
}
