% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_unit_suffix}
\alias{exp_unit_suffix}
\title{Obtain suffix for time exposure unit}
\usage{
exp_unit_suffix(exp_unit)
}
\arguments{
\item{exp_unit}{Character defining the unit of time exposure ("minutes" the
default).}
}
\value{
Character indicating the respective suffix for the \code{exp_unit}
entered.
}
\description{
Obtain suffix for time exposure unit
}
\keyword{internal}
