% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_exposures}
\alias{get_data_exposures}
\title{Extract exposures data frame}
\usage{
get_data_exposures(injd)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see
\code{\link[=prepare_all]{prepare_all()}}).}
}
\value{
The exposure data frame containing the necessary columns:
"person_id", "date" and "time_expo".
}
\description{
Extract exposures data frame from the injd object.
}
\examples{
get_data_exposures(injd)
}
