% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_followup}
\alias{get_data_followup}
\title{Extract follow-up data frame}
\usage{
get_data_followup(injd)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see
\code{\link[=prepare_all]{prepare_all()}}).}
}
\value{
The follow-up data frame containing the necessary columns:
"person_id", "t0" and "tf".
}
\description{
Extract follow-up data frame from the injd object.
}
\examples{
get_data_followup(injd)
}
