% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_injuries}
\alias{get_data_injuries}
\title{Extract injury/illness data frame}
\usage{
get_data_injuries(injd)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see
\code{\link[=prepare_all]{prepare_all()}}).}
}
\value{
The injury/illness data frame containing the necessary columns:
"person_id", "date_injured" and "date_recovered".
}
\description{
Extract injury/illness data frame from the injd object.
}
\examples{
get_data_injuries(injd)
}
