% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_injpolar.R
\name{gg_injpolar}
\alias{gg_injpolar}
\title{Polar area diagrams for available/injured players proportions}
\usage{
gg_injpolar(
  injd,
  by = c("monthly", "season"),
  var_type_injury,
  title = "Polar area diagram\\ninjured and available (healthy) players per month"
)
}
\arguments{
\item{injd}{Prepared data, an \code{injd} object.}

\item{by}{Character, one of "monthly" or "season", specifying the periodicity
according to which to calculate the proportions of available and injured
players/athletes.}

\item{var_type_injury}{Character specifying the name of the column on the
basis of which to classify the injuries and calculate proportions of the
injured players. It should refer to a (categorical) variable that describes
the "type of injury". Defaults to \code{NULL}.}

\item{title}{Text for the main title.}
}
\value{
A ggplot object (to which optionally more layers can be added).
}
\description{
Plot the proportions of available and injured players in the cohort on a
monthly or season basis. Further information of the type of injury may be
specified so that the injured players proportion is disaggregated and
reported according to this variable.
}
\examples{
# Prepare data
# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

gg_injpolar(injd, var_type_injury = "injury_type")
}
