% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_injriskmatrix.R
\name{gg_injriskmatrix}
\alias{gg_injriskmatrix}
\title{Risk matrix plot}
\usage{
gg_injriskmatrix(
  injds,
  var_type_injury = NULL,
  add_contour = TRUE,
  title = NULL,
  xlab = "Incidence (injuries per _)",
  ylab = "Burden (days lost per _)",
  errh_height = 3,
  errv_width = 0.5
)
}
\arguments{
\item{injds}{\code{injds} \strong{S3} object (see
\code{\link[=injsummary]{injsummary()}})}

\item{var_type_injury}{Character specifying the name of the column, a
(categorical) variable referring to the "type of injury" (e.g.
muscular/articular/others or overuse/not-overuse etc.) according to which
visualize injury summary statistics (optional, defaults \code{NULL}).}

\item{add_contour}{FILL}

\item{title}{Text for the main title, passed to
\code{\link[ggplot2:ggtitle]{ggplot2::ggtitle()}}.}

\item{xlab}{x-axis label to be passed to
\code{\link[ggplot2:xlab]{ggplot2::xlab()}}.}

\item{ylab}{y-axis label to be passed to
\code{\link[ggplot2:ylab]{ggplot2::ylab()}}.}

\item{errh_height}{Set the height of the horizontal interval whiskers;
\code{height} argument for
\code{\link[ggplot2:geom_errorbarh]{ggplot2::geom_errorbarh()}}}

\item{errv_width}{Set the width of the vertical interval whiskers;
\code{width} argument for
\code{\link[ggplot2:geom_errorbar]{ggplot2::geom_errorbar()}}}
}
\value{
A ggplot object (to which optionally more layers can be added).
}
\description{
Given an \code{injds} \strong{S3} object it depicts risk matrix plots (see
References section), a graph of injury severity plotted against injury
incidence. The point estimates of each measure, as well as their confidence
intervals, are plotted according to the method used when running
\code{\link[=injsummary]{injsummary()}} function.
}
\examples{
# Prepare data
# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

# Estimate injury summary statistics
injds <- injsummary(injd)
injds2 <- injsummary(injd, var_type_injury = "injury_type")

# Plot
gg_injriskmatrix(injds)
gg_injriskmatrix(injds2, var_type_injury = "injury_type", title = "Risk matrix")
}
\references{
Bahr R, Clarsen B, Derman W, et al. International Olympic Committee
consensus statement: methods for recording and reporting of epidemiological
data on injury and illness in sport 2020 (including STROBE Extension for
Sport Injury and Illness Surveillance (STROBE-SIIS)) \emph{British Journal of
Sports Medicine} 2020; 54:372-389.

Fuller, C. W. (2018). Injury Risk (Burden), Risk Matrices and Risk Contours
in Team Sports: A Review of Principles, Practices and Problems. \emph{Sports
Medicine}, 48(7), 1597–1606. https://doi.org/10.1007/s40279-018-0913-5
}
