% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injsummary.R
\name{injsummary}
\alias{injsummary}
\title{Injury summary statistics}
\usage{
injsummary(
  injd,
  var_type_injury = NULL,
  method = c("poisson", "negbin", "zinfpois", "zinfnb"),
  conf_level = 0.95,
  quiet = FALSE
)
}
\arguments{
\item{injd}{\code{injd} \strong{S3} object (see \code{\link[=prepare_all]{prepare_all()}}).}

\item{var_type_injury}{Character specifying the name of the column according
to which compute injury summary statistics. It should refer to a
(categorical) variable that describes the "type of injury". Optional,
defaults to \code{NULL}.}

\item{method}{Method to estimate injury incidence and injury burden. One of
"poisson", "negbin", "zinfpois" or "zinfnb"; characters that stand for
Poisson method, negative binomial method, zero-inflated Poisson and
zero-inflated negative binomial.}

\item{conf_level}{Confidence level (default = 0.95).}

\item{quiet}{Logical, whether or not to silence the warning messages
(defaults to \code{FALSE}).}
}
\value{
A list of two data frames comprising player-wise and overall
injury summary statistics, respectively, that constitute an \code{injds}
\strong{S3} object, and both of them made up of the following columns:
\itemize{
\item \code{ninjuries}: number of injuries sustained by the player over the
given period specified by the \code{injd} data frame.
\item \code{ndayslost}: number of days lost by the player due to injury
over the given period specified by the \code{injd} data frame.
\item \code{totalexpo}: total exposure that the player has been under risk
of sustaining an injury.
\item \code{injincidence}: injury incidence, number of injuries per unit of
exposure.
\item \code{injburden}: injury burden, number of days lost per unit of
exposure.
\item \code{var_type_injury}: only if it is specified as an argument to
function.
}
}
\description{
Calculate injury summary statistics such as injury incidence and injury
burden (REF-ak), including total number of injuries, number of days lost due
to injury, total time of exposure etc., by means of a classic-standard
(Poisson) method, negative binomial, zero-inflated poisson or zero-inflated
negative binomial, on a player and overall basis.
}
\examples{
# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

injsummary(injd)
}
