% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injsummary.R
\name{is_injds}
\alias{is_injds}
\title{Check if an object is of class \code{injds}}
\usage{
is_injds(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
A logical value: \code{TRUE} if \code{x} inherits from \code{injds}
class, \code{FALSE} otherwise.
}
\description{
Check if an object x is of class \code{injds}.
}
