% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_unit.R
\name{label_unit}
\alias{label_unit}
\title{Label the time unit}
\usage{
label_unit(unit)
}
\arguments{
\item{unit}{Character that indicates the unit of exposure of the
sports-related health problem data.}
}
\value{
A character vector that expresses the
unit used for the rates, i.e. for the athlete's time at risk, named
\code{unit_timerisk}.
}
\description{
An auxiliary function to return the label of the exposure time unit used.
}
\keyword{internal}
