% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_injd}
\alias{new_injd}
\title{constructor of injd class}
\usage{
new_injd(x = data.frame(), unit_exposure = "match_minutes")
}
\arguments{
\item{x}{a data frame object to construct it to injd class object}

\item{unit_exposure}{first attribute}
}
\value{
a new injd object
}
\description{
constructor of injd class
}
\keyword{internal}
