% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_unit.R
\name{summary_unit}
\alias{summary_unit}
\title{Transform calc_summary() output according to the unit of exposure}
\usage{
summary_unit(unit, summary_data, quiet)
}
\arguments{
\item{unit}{Character that indicates the unit of exposure of the
sports-related health problem data.}

\item{summary_data}{\code{summary_data} data frame to be transformed.}

\item{quiet}{Logical, whether or not to silence the warning messages
(defaults to \code{FALSE}).}
}
\value{
The same calc_summary() data frame with
'incidence' and 'cburden' values transformed according to \code{unit},
named \code{summary_data}.
}
\description{
Transform calc_summary() output according to the unit of exposure
}
\keyword{internal}
