% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_injd}
\alias{validate_injd}
\title{validator or injd class}
\usage{
validate_injd(x)
}
\arguments{
\item{x}{an injd class object}
}
\value{
an error if x is not of injd class; otherwise x (invisibly)
}
\description{
validator or injd class
}
\keyword{internal}
