% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date2season}
\alias{date2season}
\title{Get the season}
\usage{
date2season(date)
}
\arguments{
\item{date}{A vector of class \link[base:Dates]{Date} or
\link[base:integer]{integer}/\link[base:numeric]{numeric}. If it is
\code{integer}/\code{numeric}, it should refer to the year in which the
season started (e.g. \code{date} = 2015 to refer to the 2015/2016 season)}
}
\value{
Character specifying the respective competition season given the
date. The season (output) follows this pattern: "2005/2006".
}
\description{
Get the season given the date.
}
\examples{
date <- Sys.Date()
date2season(date)
}
