% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injprop.R
\name{injprop}
\alias{injprop}
\title{Title}
\usage{
injprop(injd, by = c("monthly", "season"), var_type_injury)
}
\arguments{
\item{injd}{Prepared data, an \code{injd} object.}

\item{by}{FILL}

\item{var_type_injury}{Character specifying the name of the column on the
basis of which to classify the injuries and calculate
proportions of the injured players. Defaults to
\code{NULL}.}
}
\value{
FILL
}
\description{
Title
}
\examples{
# df_exposures <- prepare_exp(raw_df_exposures, player = "player_name",
#                             date = "year", time_expo = "minutes_played")
# df_injuries  <- prepare_inj(raw_df_injuries, player = "player_name",
#                             date_injured = "from", date_recovered = "until")
# injd         <- prepare_all(data_exposures = df_exposures,
#                             data_injuries  = df_injuries,
#                             exp_unit = "matches_minutes")

injprop(injd, by = "monthly", var_type_injury = "injury_type")
}
