% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injurytools-package.R
\docType{package}
\name{injurytools}
\alias{injurytools-package}
\alias{injurytools}
\title{\pkg{injurytools} package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sports Injury Data analysis aims to identify and describe the magnitude of the injury problem, and to gain more insights (e.g. determine potential risk factors) by statistical modelling approaches. The 'injurytools' package provides standardized routines and utilities that simplify such analyses. It offers functions for data preparation, informative visualizations and descriptive and model-based analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lzumeta/injurytools}
  \item \url{https://lzumeta.github.io/injurytools/}
  \item Report bugs at \url{https://github.com/lzumeta/injurytools/issues}
}

}
\author{
\strong{Maintainer}: Lore Zumeta Olaskoaga \email{lorezumeta@gmail.com} (\href{https://orcid.org/0000-0001-6141-1469}{ORCID})

Other contributors:
\itemize{
  \item Dae-Jin Lee \email{dlee@bcamath.org} (\href{https://orcid.org/0000-0002-8995-8535}{ORCID}) [contributor]
}

}
\keyword{internal}
